refactor: vite-plugin-md -> vite-plugin-md-legacy
