+++ packages/astro/package.json
@@ -49,6 +49,8 @@
     "./vite-plugin-astro-postprocess/*": "./dist/vite-plugin-astro-postprocess/*",
     "./vite-plugin-jsx/*": "./dist/vite-plugin-jsx/*",
     "./vite-plugin-jsx": "./dist/vite-plugin-jsx/index.js",
+    "./vite-plugin-markdown-legacy": "./dist/vite-plugin-markdown-legacy/index.js",
+    "./vite-plugin-markdown-legacy/*": "./dist/vite-plugin-markdown-legacy/*",
     "./vite-plugin-markdown": "./dist/vite-plugin-markdown/index.js",
     "./vite-plugin-markdown/*": "./dist/vite-plugin-markdown/*",
     "./dist/jsx/*": "./dist/jsx/*"
