+++ packages/astro/src/vite-plugin-markdown-legacy/README.md
@@ -1,3 +1,3 @@
-# vite-plugin-markdown
+# vite-plugin-markdown-legacy
 
 Adds Markdown support to Vite, both at the top level as well as within `.astro` files.
