+++ packages/astro/src/vite-plugin-markdown-legacy/index.ts
