+++ packages/markdown/remark/src/remark-escape.ts
@@ -2,7 +2,7 @@ import type { Literal } from 'unist';
 import { visit } from 'unist-util-visit';
 
 // In code blocks, this removes the JS comment wrapper added to
-// HTML comments by vite-plugin-markdown.
+// HTML comments by vite-plugin-markdown-legacy.
 export default function remarkEscape() {
 	return (tree: any) => {
 		visit(tree, 'code', removeCommentWrapper);
